API: export record JSONs with site geojson
This is a first attempt for finding a suitable datastructure for representing the record data inside GeoJSON. It may change in the future.

Another change is the switch from json to simplejson library, as pythons standard implementation does not serialize statically typed Decimals to json, but simplejson does by default.

Another change fixes the loading of the record properties to the database, the json.dumps() was wrong here.